Add `customvars` to read additional hostvars from text field
This adds the possibility to read additional host vars from a text
field in GLPI. The field should be "Text (multiple lines)" and only
contain valid YAML syntax. If the syntax is invalid, the inventory
will throw an error containing the item's name and inventory number.

The text field's id is passed as the value of the new `customvars`
configuration option.

This feature introduces the PyYAML module as a new dependency.